// src/types/types.ts
export type TSlide = {
  image: any;
  bannerHeading: any;
  bannerSubHeading: any;
  text: string;
  buttonText: string;
  buttonLink: string;
};
export type TTimeRemaining = {
  days: number;
  hours: number;
  minutes: number;
  seconds: number;
};

export interface IServiceItem {
  id: number;
  icon: StaticImageData;
  title: string;
}
// ...
export interface IDiscountProducts {
  id: number;
  imageUrl: StaticImageData;
  title: string;
}